+++ lib/barrel_file_lints.dart
@@ -12,11 +12,11 @@ import 'package:barrel_file_lints/src/fixes/simplify_relative_path.dart';
 import 'package:barrel_file_lints/src/fixes/use_layer_specific_barrel.dart';
 import 'package:barrel_file_lints/src/rules/avoid_core_importing_features.dart';
 import 'package:barrel_file_lints/src/rules/avoid_cross_feature_barrel_exports.dart';
-import 'package:barrel_file_lints/src/rules/avoid_flutter_in_domain.dart';
 import 'package:barrel_file_lints/src/rules/avoid_improper_layer_import.dart';
 import 'package:barrel_file_lints/src/rules/avoid_internal_feature_imports.dart';
 import 'package:barrel_file_lints/src/rules/avoid_relative_barrel_imports.dart';
 import 'package:barrel_file_lints/src/rules/avoid_self_barrel_import.dart';
+import 'package:barrel_file_lints/src/rules/avoid_ui_framework_in_logic.dart';
 
 export 'src/fixes/convert_to_package_import.dart';
 export 'src/fixes/remove_cross_feature_export.dart';
@@ -27,11 +27,11 @@ export 'src/fixes/simplify_relative_path.dart';
 export 'src/fixes/use_layer_specific_barrel.dart';
 export 'src/rules/avoid_core_importing_features.dart';
 export 'src/rules/avoid_cross_feature_barrel_exports.dart';
-export 'src/rules/avoid_flutter_in_domain.dart';
 export 'src/rules/avoid_improper_layer_import.dart';
 export 'src/rules/avoid_internal_feature_imports.dart';
 export 'src/rules/avoid_relative_barrel_imports.dart';
 export 'src/rules/avoid_self_barrel_import.dart';
+export 'src/rules/avoid_ui_framework_in_logic.dart';
 export 'src/utils/feature_pattern_utils.dart';
 
 /// Plugin for enforcing barrel file import rules between features
@@ -47,7 +47,7 @@ class BarrelFileLintPlugin extends Plugin {
       ..registerLintRule(AvoidSelfBarrelImport())
       ..registerLintRule(AvoidCrossFeatureBarrelExports())
       ..registerLintRule(AvoidImproperLayerImport())
-      ..registerLintRule(AvoidFlutterInDomain())
+      ..registerLintRule(AvoidUiFrameworkInLogic())
       ..registerLintRule(AvoidRelativeBarrelImports())
       // Register quick fixes
       ..registerFixForRule(